Refactor to reduce size of diff
